+++ photutils/datasets/sources.py
@@ -5,16 +5,13 @@ model parameters.
 """
 
 import numpy as np
-from astropy.modeling.models import Gaussian2D
 from astropy.table import QTable
-from astropy.utils.decorators import deprecated
 
 from photutils.utils._coords import make_random_xycoords
 from photutils.utils._misc import _get_meta
 from photutils.utils._parameters import as_pair
 
-__all__ = ['make_model_params', 'make_random_models_table',
-           'make_random_gaussians_table']
+__all__ = ['make_model_params', 'make_random_models_table']
 
 
 def make_model_params(shape, n_sources, *, x_name='x_0', y_name='y_0',
@@ -160,8 +157,8 @@ def make_random_models_table(n_sources, param_ranges, seed=None):
     defined by the column names. The parameters are drawn from a uniform
     distribution over the specified input ranges, inclusively.
 
-    The output table can be input into :func:`make_model_sources_image`
-    to create an image containing the model sources.
+    The output table can be input into :func:`make_model_image` to
+    create an image containing the model sources.
 
     Parameters
     ----------
@@ -186,10 +183,6 @@ def make_random_models_table(n_sources, param_ranges, seed=None):
         are defined by the column names. The column names will be the
         keys of the dictionary ``param_ranges``. The table will also
         contain an ``'id'`` column with unique source IDs.
-
-    See Also
-    --------
-    make_random_gaussians_table
 
     Notes
     -----
@@ -228,87 +221,5 @@ def make_random_models_table(n_sources, param_ranges, seed=None):
         # Generate a column for every item in param_ranges, even if it
         # is not in the model (e.g., flux).
         sources[param_name] = rng.uniform(lower, upper, n_sources)
-
-    return sources
-
-
-@deprecated('1.13.0', alternative='make_random_models_table')
-def make_random_gaussians_table(n_sources, param_ranges, seed=None):
-    """
-    Make a `~astropy.table.QTable` containing randomly generated
-    parameters for 2D Gaussian sources.
-
-    Each row of the table corresponds to a Gaussian source whose
-    parameters are defined by the column names. The parameters are drawn
-    from a uniform distribution over the specified input ranges.
-
-    The output table will contain columns for both the Gaussian
-    amplitude and flux.
-
-    The output table can be input into
-    :func:`make_gaussian_sources_image` to create an image containing
-    the 2D Gaussian sources.
-
-    Parameters
-    ----------
-    n_sources : float
-        The number of random 2D Gaussian sources to generate.
-
-    param_ranges : dict
-        The lower and upper boundaries for each of the
-        `~astropy.modeling.functional_models.Gaussian2D` parameters
-        as a dictionary mapping the parameter name to its ``(lower,
-        upper)`` bounds. The dictionary keys must be valid
-        `~astropy.modeling.functional_models.Gaussian2D` parameter
-        names or ``'flux'``. If ``'flux'`` is specified, but not
-        ``'amplitude'`` then the 2D Gaussian amplitudes will be
-        calculated and placed in the output table. If ``'amplitude'``
-        is specified, then the 2D Gaussian fluxes will be calculated
-        and placed in the output table. If both ``'flux'`` and
-        ``'amplitude'`` are specified, then ``'flux'`` will be
-        recalculated and overwritten. Model parameters not defined in
-        ``param_ranges`` will be set to the default value.
-
-    seed : int, optional
-        A seed to initialize the `numpy.random.BitGenerator`. If `None`,
-        then fresh, unpredictable entropy will be pulled from the OS.
-
-    Returns
-    -------
-    table : `~astropy.table.QTable`
-        A table of parameters for the randomly generated Gaussian
-        sources. Each row of the table corresponds to a Gaussian source
-        whose parameters are defined by the column names.
-
-    See Also
-    --------
-    make_random_models_table
-
-    Notes
-    -----
-    To generate identical parameter values from separate function
-    calls, ``param_ranges`` must have the same parameter ranges and the
-    ``seed`` must be the same.
-    """
-    sources = make_random_models_table(n_sources, param_ranges,
-                                       seed=seed)
-    model = Gaussian2D()
-
-    # compute Gaussian2D amplitude to flux conversion factor
-    if 'x_stddev' in sources.colnames:
-        xstd = sources['x_stddev']
-    else:
-        xstd = model.x_stddev.value  # default
-    if 'y_stddev' in sources.colnames:
-        ystd = sources['y_stddev']
-    else:
-        ystd = model.y_stddev.value  # default
-    gaussian_amplitude_to_flux = 2.0 * np.pi * xstd * ystd
-
-    if 'amplitude' in param_ranges:
-        sources['flux'] = sources['amplitude'] * gaussian_amplitude_to_flux
-
-    if 'flux' in param_ranges and 'amplitude' not in param_ranges:
-        sources['amplitude'] = sources['flux'] / gaussian_amplitude_to_flux
 
     return sources